hack/setup/scripts/helm-converter/requirements.txt
# Requirements for Kustomize to Helm Converter
# All other dependencies are Python standard library

# YAML parsing and generation
PyYAML>=6.0

# Testing
pytest>=7.0
